web-app/requirements.txt
Flask==2.3.3
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation==0.41b0
Werkzeug==2.3.7
